Rename to zuulcilint
PyPi considers zuullint similar to zuul-lint and the new
package is not allowed.
