+++ .gitignore
@@ -130,6 +130,3 @@ dmypy.json
 
 # ruff
 .ruff_cache/
-
-# Generated by setuptools-scm
-zuullint/_version.py
